[rpmb_dev] Convert rpmb_dev to a daemon
Previously, rpmb_dev would take a file on the command line to read a
single complete request, and write a single complete response to.
This function has been replaced by operating as a daemon on a unix
domain socket at a provided path.

The --init functionality is unchanged and not daemonized.

Bug: 124277696
Change-Id: I83a831988ce5ae9d9874f0da8a9688f8a8c46b9c