rpmb_dev: Suppress error messages already reported
rpmb_dev currently prints any error condition to stderr, including those
already reported to the client.

Since Trusty probes the size of the rpmb device through out of bounds
read requests, this clutters the log. Trusty also receives an error code
in all of these cases, allowing it to make the decision of whether to
print a message. Error messages which "shouldn't happen" (e.g. specific
to the implementation of rpmb_dev, or which are not otherwise reported
to the client) are still printed.

The old behavior (printing all errors) can be accessed by passing
--verbose to rpmb_dev.

Bug: 133358996
Change-Id: Ia644282b5bd4f07bf4c1bb7100859e0df2673c75