storage: block_cache: Use new obj_ref_add_allow_unreferenced_obj
To reduce errors in the usage of refs, obj_ref_add now checks that there
is at least one reference to the object when creating it. The block
cache intentionally runs with 0 reference count in order to recycle
existing entries in the block cache. obj_ref_add_allow_unreferenced_obj
skips this check, preventing this code from breaking.

Bug: 117221195
Change-Id: I5792ef06fa1df12158fa78261116b87eccfaf668