storage: Support RPMB key provisioning
If the RPMB device has not been programmed with a RPMB key,
a 32 bytes random number is generated. The number is used to
derive a RPMB key with which the RPMB device is programmed.
The random number is written into the first block of the device.

For the situation where the RPMB device has been programmed,
the 32 bytes random number is read from the device.
The RPMB key can be re-derived with the random number.

Bug: 152901318
Change-Id: I91487bfdacb66d18081088b657ba23f4f6e67fd8