storage: block_cache: Fix assert in block_cache_discard_transaction
Change assert of dev field in block cache entry to allow dirty tmp
blocks block to be on a separate device then dirty non-tmp blocks. If a
write of the super block fails on the td filesystem, the superblock will
be dirty when we try to discard the transaction. Since there could also
be dirty tmp blocks at this point the assert that makes sure all the
dirty blocks are on the same device would fail. Dirty non-tmp blocks
have to be cleaned before the superblock is written so we keep the
existing assert for non-tmp blocks.

Currently this change does not prevent the storage server from going
down as there is also an assert in the transaction layer that the
superblock write must not fail, but since that assert is more
informative it is better to abort there than in the block cache layer.
Aslo, if we want to add a recovery path in the future, discard will need
to work.

Bug: 171886785
Change-Id: Idb8cf05ed7e28697708a59c8901c8536d8bcc922