storage: Add a clear superblock flag and use it
After a user-data wipe we detect that the non secure backing file for
the td filesystem is empty and we create an in-memory emoty filesystem
to match instead of loading the old superblock from the rpmb partition.
When we start writing to this new filesystem we need to write to the
backing file before we can write a valid superblock. A reboot at this
point will cause the old superblock to be loaded again as the backing
file is no longer empty. To avid this problem, we add a new superblock
flag that indicates that the filesystem is empty (there is no root node
in the file and free trees) and write this before we write any data to
the backing file.

Bug: 190109878
Change-Id: I87366ecc526a04174b7dd278b2bf27fbe4854e18