storage: Discard all cache entries for a block device on teardown
When the non-secure storage proxy disonnects we need to also discard all
cache entries that point to the block device we created for it. Without
this we have blocks in the cache that points to a memory object that has
been freed and allocated again. With the current heap, the object
happens to be allocated at the same address and will match the new
device created afer a re-connect. This makes it difficult to test how
the filesystem behaves after a reboot by just restarting the proxy.

Bug: 190109878
Change-Id: Ic4ee8ebcfa8b4c54bc7d16ee60dbc723546ce3a5