Fix discarding dirty, referenced block
block_put_dirty_discard should accept a referenced block and drop the
reference then discard the block data. Instead it was discarding
the data first, triggering an assertion in
block_cache_entry_discard_dirty that expected the discard entry to not
have a dirty_ref. Reverses the ordering so that the function first drops
the reference to the block, then discards the (potentially) dirty block.

Adds a tests that triggers this assert before the fix by attempting to
discard a dirty, referenced block.

Test: build.py generic-arm64-test-debug --test storage_test
Bug: 199726432
Change-Id: I3e3d196d8aa838c43b06ad314f99a2259f7bd4c0