Resync write counter on write failure
If the RPMB chip performs a write but we fail to validate the response,
our view of the write counter state will be off by one and future writes
will fail until the device is rebooted and the filesystem state is
re-initialized. Instead we resync the write counter by reading it from
the chip (an authenticated operation with replay protection), allowing
us to rewrite the current, valid superblock and recover to a known
state.

Test: build.py generic-arm64-test-debug --test storage_test
Bug: 198638362
Bug: 200187979
Change-Id: Ic4f1c4f026384357ce10580821cab6fbfd44344b