Recover FS if superblock rewrite fails
Failing a superblock rewrite, either because the initial cleared
superblock failed to commit or because a forced rewrite of the current
state failed to commit, would previously cause the FS to go into a
read-only state until re-initialization. This change adds the ability to
recover from this state by resetting the transaction and trying the same
write again.

Test: build.py qemu-generic-arm64-test-debug --test storage-tp-clear-test
Bug: 202792882
Bug: 198638362
Bug: 194313068
Change-Id: I06016f786b228fd30d7923b7303bc1e0e960a52e