Update rpmb_dev license to Apache 2.0
rpmb_dev is a rpmb device stub used in emulation
as well as platform early bringup so we don't expect
any open source developer to be impacted by the migration
from MIT to Apache 2.0.
Note that with such a migration to Apache 2.0, recipients
won't receive lesser permissions, they just have updated
requirements for which license text to share along with the code.

Bug: 191508826
Change-Id: Ibf2aa7bc4a7d7ea27029d6cbb09acae34dc60f83