storage: Support switching between two superblocks
Swaps the current superblock with a backup when the storage TA detects
that the normal backing file location for a filesystem is inaccessible
but an alternate data file is available. This indicates that the device
is in an "alternate data" mode (i.e. a dynamic system update image is
running). We store the current alternate data state in the superblock,
which allows us to detect when the state changes and therefore when the
superblock should be swapped with its backup. When we detect that the
current superblock does not match the expected mode we either swap the
current superblock roots with those in the backup slot, or we back up
and clear the roots (depending on whether we detect that we need to
clear the fs). This functionality is gated on the
STORAGE_NS_ALTERNATE_SUPERBLOCK_ALLOWED build flag.

This backup and restore allows the TD filesystem to be temporarily
cleared, e.g. during a dynamic system update (DSU) using a Generic
System Image (GSI). A DSU temporarily mounts a new userdata partition,
thus clearing the TD filesystem. However, the DSU is temporary and when
the device reverts back to its original installation and userdata
partition, Trusty storage would previously detect this as a rollback if
any writes occurred while in DSU mode. With this new functionality,
storage will back up the superblock upon entering DSU mode and detecting
a missing TD backing file. When DSU mode is turned off, storage will
attempt to swap the backup (which corresponds to the original TD backing
file) with the current superblock (which was for GSI) if it matches the
expected backing file state. We back up rather than clear in this case
because DSU allows rebooting into and out of the same GSI with the same
userdata partition multiple times.

Test: build.py generic-arm64-test-debug --test storage_test
Bug: 203719297
Change-Id: I45137264bb1cdc17d92aefeb3f56864b6c028fb8