Add IPC port to manifest
If the port for an application is not available, ql-tipc
immediately returns with an error without waiting for the port
unless that port is listed in the manifest. This adds the storage
port to the manifest to fix a race between the bootloader
connecting to storage and the app starting during boot.

Bug: 217580763
Change-Id: I8318cac6d44f968f406d10341181e9f706b811f3