storage: Add nonsecure persistent FS
Adds a new filesystem port which provides confidential, persistent, but not
rollback-protected, storage. This filesystem may attempt to recover from
potential corruption by deleting files or wiping itself entirely, with
the goal of not getting stuck in an inoperable state.

If possible, use a different filesystem that provides tamper-detection.
This filesystem is intended as a safe alternative to TDP when writing
onto unreliable backing stores.

Bug: 228793975
Test: com.android.storage-unittest.nsp
Change-Id: I3491594f7a173eac46dee8aa3b0032f131d2aa64