storage: Add filesystem check and repair functionality
If the storage filesystem becomes corrupted due to issues in the
nonsecure backing file, we generally do not allow modification of that
filesystem as it has been "tampered" with. In some cases we want to
allow a filesystem to repair itself. We previously only supported repair
by wiping the entire filesystem. This change adds the capability to
delete individual files that are corrupted while preserving the rest of
the filesystem that is still valid.

To support early detection and optional repair, this change also adds a
filesystem check that iterates all files in the filesystem and attempts
to fetch at least one data block from each to ensure that the metadata
for each file is valid.

Bug: 228793975
Test: com.android.storage-unittest.nsp
Change-Id: I73fee6b746a2f76a6906bf5a0b1a0f221e7a652c