storage: Add superblock flag to block cache entries
Previously we had no way to know when writing back a block entry whether
it was for a superblock and thus necessitates a file sync to ensure data
is flushed before committing the transaction. Adds a flag on block cache
entries to indicate if they are superblocks. Only new blocks for writing
a superblock get this flag currently, but that's all that matters since
we only care about this when flushing writes.

Test: build.py --test com.android.storage-unittest.td
Bug: 228793975
Change-Id: I9cb62875b7165df02f09afdc2243e9ae4b6e7378