storage: Fail all pending transactions if fsync fails
When the storage proxy reports an fsync failure, we don't know which
transaction the failing write belongs to. A prior write that returned
successfully might have a delayed failure. To be safe, we need to fail
all transactions that might have touched the file that failed to fsync.
Because we don't track the failing file yet, fail all pending
transactions.

Test: manually inject errors to cause an fsync failure for a pending transaction
Bug: 239105007
Change-Id: I735171b2221f82ca45e12117335eb5e14fcf76a4