storage: request reboot after init tests
Some storage tests must be separated by a reboot in order to check
that data was correctly persisted. Use the compositetest element to
capture the ordering requirement between the init, check, and clean
steps of the storage tests for td and tdea. Use reboot command to
request reboots when sharing QEMU instances among androidtests.

Bug: 242077104
Change-Id: Ied1c1ee7d0a9acab1c1e379b1afe0dc0a4d8dd87