storage: Trusty-side IPC to obtain STORAGE_FILE_GET_MAX_SIZE
go/oag/2241573 defined the STORAGE_FILE_GET_MAX_SIZE API on the
Trusty-side. This CL adds the IPC call to Android (and also
adds it to the storageproxy_shim used for tests).
(The Android-side changes to implement STORAGE_FILE_GET_MAX_SIZE,
and changes to block_device_tipc.c to use the max size, are left
as an exercise for the reader.)

Bug: 247013568
Test: None
Change-Id: I6d66da266bd5e5325206aacbe41dc996b24c8ddf