storage: Add checkpoint to super block and file-system
We want to support checkpointing the file-system state to allow
controlled rollback in the event of file-system corruption. This
checkpoint is a file tree, just like the normal file-system tree, but
all blocks in the checkpoint will be left in place even if these blocks
are (copy-on-write) modified in the file-system.

Adds a checkpoint root node to the super block (and its backup
structure). This checkpoint node will point to a block containing a
pointer to the checkpoint file tree root and pointer to the root of a
block set containing the free blocks at the time of the checkpoint. The
block set allows quick checking of blocks during allocation so that we
can avoid reusing blocks in the checkpoint.

Test: None
Bug: 244607290
Change-Id: Iad7f186c898957ee17eb2e70930f6338c3d911c7