storage: don't rely on nonstandard realloc() behavior
The result of realloc(ptr, 0) is implementation-specific, and therefore
calling realloc() with a size of zero should be avoided. Specifically,
scudo will free the pointer and return NULL, whereas dlmalloc will free
the pointer and return a new non-NULL pointer.
The storage app was relying on dlmalloc's behavior and broke with scudo.
Fix by avoiding calling realloc with a size of 0.

Bug: None
Test: run storage app with scudo

Change-Id: I28f248f2ee9c288d271a696f4600f470243a42d6