storage: Add a repaired flag to the superblock
Adds a repaired flag to the superblock to track whether any repairs have
been made to a filesystem. Repairs are equivalent to partial FS
rollback, so we must notify clients that data may have been rolled back
after a repair has occurred. Older software versions that do not
recognize and preserve this repaired flag will be unable to mount a
filesystem in the repaired state, preserving the state to ensure clients
are notified. Older software that does not preserve the repair state may
mount the filesystem after it has been cleared and the flag reset,
however.

Test: storage_block_test
Bug: 244608684
Change-Id: I367179ea47efb71c99dfaac57c158d8f6b68430d