storage: Return more descriptive results from file_open
The file_open() API returned a bool but could fail to open a file for
various reasons that need to be distinguished. This changes adds a new
result enum and propagates errors from file_open() in a more consistent
way.

Attempting to open a file that is already open in the same transaction
now results in STORAGE_ERR_NOT_ALLOWED rather than
STORAGE_ERR_NOT_FOUND, and the test case has been updated accordingly.
This should have been the behavior originally but the information to
determine this case was not available to the TIPC service layer.

Test: com.android.storage-unittest.*
Test: storage_block_test
Test: storage_host_test
Bug: 261112283
Change-Id: I5e7ea9430a52b458fe01b9d5abdeaeba2f8840ff