storage: super: Split fs_version into version and flags
Adds a fs_flags field to the super block by shrinking the fs_version
field down to 16 bits from 32 bits and using the remaining 16 bits for
the new fs_flags field. These fs_flags will be used to describe features
that are required to be supported to mount or modify the super block.
This allows the addition of (toggleable) features that are not backwards
compatible without requiring a new filesystem version.

Test: storage_block_test
Bug: 244608684
Change-Id: I8bab422872e6da23706fe5e6143ece658e5470bb