storage: transaction: Detect and propagate block errors
Detect and distinguish MAC mismatch and missing block errors at the
block layer, and note these errors in the transaction state. This allows
higher layers to determine if a block read failure could be transient or
was definitely due to corrupted or missing data.

Also improves block_probe to not return a failure unless we positively
confirm an invalid block.

Test: storage_block_test
Bug: 244608684
Change-Id: I5dc60949367e958b82ea495eb43d7921a3a460af