storage: fs: Do not check data blocks in fs_check
Removes data block checks when checking the file-system. Data block
corruption can be handled on a per-file basis, we just need to check the
internal tree consistency when doing a full file-system check.

Also removes the implementation of deleting a file if corrupted blocks
are detected in the file. We were not using this implementation and
callers should be responsible for implementing necessary repairs, rather
than mixing check and repair operations.

Test: storage_block_test
Bug: 244608684
Change-Id: Ic5d0ccf7b04b06410b6c439466a951a8e12fa6d5