storage: Do not crash if RPMB verification fails
To verify that the first RPMB write after a reboot or error was
successful, we read the block contents back. We would immediately crash
if this verification fails, which may occurr if the proxy is dead
because the device is shutting down. To avoid this spurious Trusty crash
during shutdown, this change instead notes that verification failed and
will not allow any further RPMB data read or write operations until the
filesystem is remounted (i.e. reboot or storageproxy restarted).

Test: storage_host_test
Bug: 263519671
Change-Id: Ifa47e662796a400b2838d4c76b45bb519ab6946d