storage: Mark FS for full scan on invalid blocks
We don't want to fully scan the entire file system for errors on each
mount. As a compromise, we check that the files and free root are
accessible with the expected MAC each time the file system is mounted.
If invalid or missing blocks are encountered later when using a file
system, we want to note this state in the super block so that we can do
a full scan and potentially repair when the file system is next mounted.
This change adds support for this flag, sets it when relevant, and
checks the flag on mount.

Test: storage_block_test
Bug: 244608684
Change-Id: I6aa56855494e68ec89ace0a2973809fb471ab6e3