storage: Allow auto-checkpointing after provisioning
If a checkpoint was not created for a filesystem during provisioning, we
may still want to create a baseline checkpoint on a subsequent boot
after this change. This change adds FS_INIT_FLAGS_AUTO_CHECKPOINT to
indicate that the filesystem initialization should attempt to save the
current filesystem state to a checkpoint if one does not already exist.
This newly created checkpoint state will then be committed as part of
the next transaction to commit to that filesystem.

This change enables auto-checkpointing for the TDP filesystem if the
build system flag STORAGE_TDP_AUTO_CHECKPOINT_ENABLED is true.

Test: com.android.storage-unittest.tdp
Bug: 269180479
Change-Id: I84895579022aeef0e459fe25c4c25a107ac8f7b1