storage: client_tipc: Checkpoint without active transaction
STORAGE_END_TRANSACTION with STORAGE_MSG_FLAG_TRANSACT_COMPLETE is
normally a no-op, but adding the STORAGE_MSG_FLAG_TRANSACT_CHECKPOINT
flag indicates that the client wants to create a checkpoint of the
current state. This change allows that combination to create a
checkpoint even when a checkpoint did not exist previously.

Test: manual
Bug: 244611675
Change-Id: I0422996279f958423fe85fb830fd2ab4cda818bf