storage: tipc_ns: Return number of bytes read
Previously ns_read_pos would either return the requested number of
bytes, if the read was completely successful, or a negative error code.
We sometimes want to know the exact number of bytes, e.g. to determine
if a read attempt was past the end of the file. This change removes the
early return of ERR_NOT_VALID if fewer than the requested number of
bytes were received.

Test: storage-td-corrupt-test (not in tree yet)
Bug: 244608684
Change-Id: I9cd75c1c464dd4e7699a152735e85d2158269596