storage: Fix TIPC timeout logging
Trusty log messages must be terminated with a newline. This change adds
missing newlines to the wait timeout logging for storage to storageproxy
TIPC communication.

Trusty masks large integers in release builds to prevent accidentaly
leakage of sensitive values. The time elapsed in milliseconds should
always be printed; this is not a sensitive value. This change also adds
the custom "x" modifier to ensure these time elapsed values are always
printed.

Test: manual
Bug: 270197433
Change-Id: I619e429bcc2cafe00bf9dbea6edea5ab0ef922e9