storage: Add repair ack flag to more file operations
Adds an allow_repaired flag to file_{delete,move,iterate} to indicate
whether the operation can proceed even if the filesystem has been
repaired. These operations previously did not check the repair status.
Prohibits all file operations on a repaired filesystem without this
flag, because the filesystem may have been rolled back by the repair
which is a change to the security semantics of the filesystem. Clients
must expliticly accept this possibility if they wish to continue
accessing files after a filesystem-wide repair.

Test: storage_block_test
Test: storage_host_test
Bug: 244608684
Change-Id: I4a645cc7cde7329bc4b86fe8817c07cf9e02f005