storage: Add metrics testing to storage_block_test
Adds testing for storage error reporting functionality to the
storage_block_test host test. We can't directly use the metrics atoms,
as the build system module type used to compile those auto-generated
sources is not compatible with a host test. Instead, we mock the
relevant APIs to track reported errors for the host test.

Test: storage_block_test
Bug: 235276465
Change-Id: I568871d9f3d8b91db1d782ac7dff032d61b8e1b1