storage: Do not probe invalid files root block
The files root block mac may be zero if the filesystem was cleared, then
an empty transaction is completed. In this case the superblock is not
"clear" and the free set is valid, but the files tree root block mac
will still be zero, i.e. invalid. We don't consider this case an error,
so this change adds support to avoid a probe failure for this case.

Test: storage_block_test and inspect for "Backing file probe failed"
Bug: None
Change-Id: I09fbf587b10a03d1d72f06763638da9dcf5eba7d