storage: Add writable flag to filesystem
Adds a writable flag to track if a filesystem was correctly initialized
and modification should be allowed. This flag defaults to false, i.e.
read-only. If errors occur during initialization of a file system, we
attempt to initialize the file system as much as possible, ensuring a
valid state, and mark the file system as not writable. This allows
initialization of other file systems to continue.

Test: storage_block_test
Bug: 273832845
Bug: 272967596
Change-Id: I2115e3227229a17fdbb17fc0aeb98ce2a4b9d8d9