storage: Add fs variants to block_mac API
The block_mac API required a transaction to access filesystem
parameters, but did not actually use the transaction itself for
anything. We would like to be able to query static properties about a
block mac, e.g. is it valid, without needing to instantiate a fake
transaction when we only have access to a filesystem structure. This
change adds block_mac_valid_fs() and block_mac_to_block_fs() which do
not require a transaction but directly use the filesystem itself.

Test: build.py
Bug: None
Change-Id: If4c8ba32d3c3c470b3fb4c5fcaf30ff66a0b6896