storage: Do not clear FS before restoring checkpoint
A previous change added a call to write_initial_super_block() when
restoring a checkpoint, but this function unconditionally clears the FS,
rather than writing the current state. We instead must call
write_current_super_block() to ensure that the restored state is written
before we overwrite any previously in-use data blocks.

Test: storage_block_test
Bug: 244608684
Change-Id: I6912e6d05cf379dedcff3583635924f0a94d9c14