storage: update scudo config to work with upstream changes
This now uses the upstream USE_CUSTOM_SCUDO_CONFIG define
and moves the Trusty size class from size_class_map.h to
custom_scudo_config.h, and eliminates the majority of the
rest of our plumbing.

We still need wrappers.cpp because the build system needs
to build a separate .o file for the files that that .cpp
includes, and otherwise shares the rest of scudo's object
files between all scudo consumers.

Bug: 231151082
Test: make with TRUSTY_APP_ALLOCATOR_FOR_trusty/user/app/storage=trusty/user/app/storage/scudo/

Change-Id: Ifde5ca9d4008d55441c7f0bab48f0d632e64df72